Changed Price and Rating to DECIMAL

+++ WolfieShop/shop/models.py
@@ -5,7 +5,7 @@ class Item(models.Model):
     itemid = models.AutoField(db_column='ItemId', primary_key=True)
     itemname = models.CharField(db_column='ItemName', max_length=100, blank=False, null=False)
     quantity = models.IntegerField(db_column='Quantity', blank=False, null=False)
-    price = models.IntegerField(db_column='Price', blank=False, null=False)
+    price = models.DecimalField(max_digits=10,decimal_places=2,db_column='Price', blank=False, null=False)
 
     CATEGORIES = (
 		('BOOKS', 'Books'),
@@ -20,7 +20,7 @@ class Item(models.Model):
     	null=False
     )
     
-    rating = models.IntegerField(db_column='Rating', blank=True, null=True, default=0)
+    rating = models.DecimalField(max_digits=3,decimal_places=2,db_column='Rating', blank=True, null=True, default=0)
     numreviews = models.IntegerField(db_column='NumReviews', blank=False, null=False, default=0) 
 
     class Meta:
